==> Extended_Figure6/02_Flow_scRNA_Frequency_Comparison/02_Get_scRNA_Freq.ipynb <==
suppressPackageStartupMessages(library(dplyr))
suppressPackageStartupMessages(library(tidyr))
suppressPackageStartupMessages(library(tibble))
suppressPackageStartupMessages(library(stringr))
library(reshape2)
library(data.table)
library(parallel)
library(purrr)
source('/home//jupyter/BRI_Figures_Final_V2/helper_function/helper_function_IHA.r')

# Read meta data and subset matched sample

meta_data<-read.csv("/home//jupyter/BRI_Figures_Final_V2/Extended-Figure4/01_Frequency_Comparison/selected_samples_with_acutal_flu_year.csv")

donors <- c("BR1025","BR1031","BR1032","BR1047","BR1050","BR1057",
            "BR2013","BR2021","BR2023","BR2039","BR2044","BR2051")

meta_data_filtered<-meta_data %>% filter(subject.subjectGuid  %in% donors ,Flu_Year=='2020-2021',Flu_Day %in% c('Day 0','Day 7'))

# Read Freq Files and do subsetting

df_freq<-read.csv("/home/jupyter/BRI_Figures_Final_V2/Dataset/diha_AIFI_L3_frequencies_2024-05-05.csv")

df_freq_Y1<-df_freq %>% filter(specimen.specimenGuid %in% meta_data_filtered$pbmc_sample_id)

df_freq_Y1<-left_join(df_freq_Y1,meta_data_filtered[c('pbmc_sample_id','Flu_Year','Flu_Day')],by=c("specimen.specimenGuid"="pbmc_sample_id"))

write.csv(df_freq_Y1,"freq_actual_flu_year_1_D0_D7.csv")
